+++ src/main/java/com/withus/withmebe/member/service/AuthService.java
@@ -6,8 +6,8 @@ import static com.withus.withmebe.common.exception.ExceptionCode.PASSWORD_CHK_MI
 import static com.withus.withmebe.common.exception.ExceptionCode.PASSWORD_MISMATCH;
 
 import com.withus.withmebe.common.exception.CustomException;
-import com.withus.withmebe.member.dto.SigninDto;
-import com.withus.withmebe.member.dto.SignupDto;
+import com.withus.withmebe.member.dto.auth.SigninDto;
+import com.withus.withmebe.member.dto.auth.SignupDto;
 import com.withus.withmebe.member.entity.Member;
 import com.withus.withmebe.member.repository.MemberRepository;
 import com.withus.withmebe.security.jwt.provider.TokenProvider;
